+++ R/CI.R
@@ -38,6 +38,7 @@
 CI = function(x, k, na.rm = FALSE, na_flag = -9999) {
 
   # Checks
+  if(inherits(x, "stars_proxy")) stop("'x' must be 'stars', not 'stars_proxy'")
   x = check_one_attribute(x)
   x = check_2d(x)
   stopifnot(is.numeric(k))