+++ R/slope.R
@@ -43,6 +43,7 @@
 slope = function(x, na_flag = -9999) {
 
   # Checks
+  if(inherits(x, "stars_proxy")) stop("'x' must be 'stars', not 'stars_proxy'")
   x = check_one_attribute(x)
   x = check_2d(x)
   stopifnot(is.numeric(na_flag))